requirements.txt
requests>=2.31,<3
Pillow>=10.0,<11
boto3>=1.34,<2
playwright>=1.44,<2
